server/rz_mesh: fix #637, a gevent/gevent-socketio issue
The actual issue is that gevent-socketio adds a header of
Access-Control-Max-Age and uses a python int instead of a string (3600),
plus the fact that gevent is pedantic about the type and doesn't call
str. So instead of fixing it on either project, we workaround in rz_mesh
(a single line).